+++ docs-website/config/navigation.tsx
@@ -56,8 +56,16 @@ const navigation = [
 						href: '/docs/getting-started/remix-quickstart',
 					},
 					{
+						title: 'Relay',
+						href: '/docs/getting-started/relay-quickstart',
+					},
+					{
 						title: 'SvelteKit',
 						href: '/docs/getting-started/sveltekit-quickstart',
+					},
+					{
+						title: 'Nuxt',
+						href: '/docs/getting-started/nuxt-quickstart',
 					},
 				],
 			},
@@ -94,12 +102,20 @@ const navigation = [
 						href: '/docs/examples/nextjs-react-query',
 					},
 					{
-						title: 'SvelteKit + SSR with TanStack Query',
-						href: '/docs/examples/sveltekit',
+						title: 'Next.js + Relay',
+						href: '/docs/examples/nextjs-relay',
 					},
 					{
 						title: 'Next.js + Clerk',
 						href: '/docs/examples/nextjs-clerk',
+					},
+					{
+						title: 'SvelteKit + SSR with TanStack Query',
+						href: '/docs/examples/sveltekit',
+					},
+					{
+						title: 'Astro',
+						href: '/docs/examples/astro',
 					},
 					{
 						title: 'Vite + SWR',
@@ -108,6 +124,10 @@ const navigation = [
 					{
 						title: 'Vite + Solid.js',
 						href: '/docs/examples/vite-solidjs',
+					},
+					{
+						title: 'Vite + React Relay',
+						href: '/docs/examples/vite-react-relay',
 					},
 					{
 						title: 'Vite + Svelte',
@@ -192,6 +212,106 @@ const navigation = [
 					{
 						title: 'WunderGraph open telemetry',
 						href: '/docs/examples/open-telemetry',
+					},
+					{
+						title: 'Nuxt + Vue Query',
+						href: '/docs/examples/nuxt',
+					},
+				],
+			},
+		],
+	},
+	{
+		title: 'About WunderGraph',
+		href: '/manifesto',
+		icon: <QuestionMarkCircleIcon />,
+		paths: ['/docs/frequently-asked-questions', '/usage-information', '/about', '/docs/use-cases'],
+		links: [
+			{
+				title: 'Manifesto',
+				href: '/manifesto',
+			},
+			{
+				title: 'Usage Information',
+				href: '/usage-information',
+			},
+			{
+				title: 'WunderGraph for',
+				links: [
+					{
+						title: 'Frontend Developers',
+						href: '/about/for-frontend-developers',
+					},
+					{
+						title: 'Fullstack Developers',
+						href: '/about/for-fullstack-developers',
+					},
+					{
+						title: 'Backend Developers',
+						href: '/about/for-backend-developers',
+					},
+				],
+			},
+			{
+				title: 'Use Cases',
+				links: [
+					{
+						title: 'Overview',
+						navTitle: 'Use Cases',
+						href: '/docs/use-cases',
+					},
+					{
+						title: 'Programmable API Gateway',
+						href: '/docs/use-cases/programmable-api-gateway',
+					},
+					{
+						title: 'API Management',
+						href: '/docs/use-cases/api-management',
+					},
+					{
+						title: 'Backend for Frontend',
+						href: '/docs/use-cases/backend-for-frontend',
+					},
+					{
+						title: 'API Composition & Integration',
+						href: '/docs/use-cases/api-composition-and-integration',
+					},
+					{
+						title: 'Versionless APIs: Easily build backwards compatible APIs',
+						href: '/docs/use-cases/versionless-apis-easily-build-backwards-compatible-apis',
+					},
+					{
+						title: 'Multi Database, multi schema ORM',
+						href: '/docs/use-cases/multi-database-multi-schema-orm',
+					},
+					{
+						title: 'Generate SDKs for all your APIs',
+						href: '/docs/use-cases/generate-sdks-for-all-your-apis',
+					},
+					{
+						title: 'Centralized Governance, Monitoring, Access Controls and Logging across your APIs',
+						href: '/docs/use-cases/centralized-governance-monitoring-access-controls-and-logging-across-your-apis',
+					},
+					{
+						title: 'Enabling your Organization to become API first',
+						href: '/docs/use-cases/enabling-your-organization-to-become-api-first',
+					},
+				],
+			},
+			{
+				title: 'Frequently Asked Questions',
+				links: [
+					{
+						title: 'How is WunderGraph faster and more secure than other GraphQL solutions?',
+						href: '/docs/frequently-asked-questions/how-is-wundergraph-faster-and-more-secure-than-other-graphql-solutions',
+					},
+					{
+						title: 'How is the developer experience different from legacy GraphQL?',
+						href: '/docs/frequently-asked-questions/how-is-the-developer-experience-different-from-legacy-graphql',
+					},
+					{
+						title: 'How is server-side only GraphQL different from client-side GraphQL?',
+						href: '/docs/frequently-asked-questions/how-is-server-side-only-graphql-different-from-client-side-graphql',
 					},
 				],
 			},
@@ -344,11 +464,6 @@ const navigation = [
 				title: 'Cookie-based Auth',
 				links: [
 					{
-						title: 'Overview',
-						navTitle: 'Cookie-based Auth',
-						href: '/docs/auth/cookie-based-auth',
-					},
-					{
 						title: 'OpenID Connect',
 						href: '/docs/auth/cookie-based-auth/openid-connect',
 					},
@@ -374,17 +489,16 @@ const navigation = [
 				title: 'Token-based Auth',
 				links: [
 					{
-						title: 'Overview',
-						navTitle: 'Token-based Auth',
-						href: '/docs/auth/token-based-auth',
+						title: 'OpenID Connect JWKS',
+						href: '/docs/auth/token-based-auth/openid-connect',
 					},
 					{
 						title: 'Auth.js (NextAuth)',
 						href: '/docs/auth/token-based-auth/auth-js',
 					},
 					{
-						title: 'OpenID Connect',
-						href: '/docs/auth/token-based-auth/openid-connect',
+						title: 'Clerk.com',
+						href: '/docs/auth/token-based-auth/clerk',
 					},
 				],
 			},
@@ -675,6 +789,10 @@ const navigation = [
 						href: '/docs/supported-frontend-frameworks/vue',
 					},
 					{
+						title: 'Nuxt',
+						href: '/docs/supported-frontend-frameworks/nuxt',
+					},
+					{
 						title: 'SolidJS',
 						href: '/docs/supported-frontend-frameworks/solidjs',
 					},
@@ -700,54 +818,6 @@ const navigation = [
 						href: '/docs/supported-backend-languages-frameworks/java-kotlin',
 					},
 				],
-			},
-		],
-	},
-	{
-		title: 'Use Cases',
-		href: '/docs/use-cases',
-		icon: <LightBulbIcon />,
-		links: [
-			{
-				title: 'Overview',
-				navTitle: 'Use Cases',
-				href: '/docs/use-cases',
-			},
-			{
-				title: 'Programmable API Gateway',
-				href: '/docs/use-cases/programmable-api-gateway',
-			},
-			{
-				title: 'API Management',
-				href: '/docs/use-cases/api-management',
-			},
-			{
-				title: 'Backend for Frontend',
-				href: '/docs/use-cases/backend-for-frontend',
-			},
-			{
-				title: 'API Composition & Integration',
-				href: '/docs/use-cases/api-composition-and-integration',
-			},
-			{
-				title: 'Versionless APIs: Easily build backwards compatible APIs',
-				href: '/docs/use-cases/versionless-apis-easily-build-backwards-compatible-apis',
-			},
-			{
-				title: 'Multi Database, multi schema ORM',
-				href: '/docs/use-cases/multi-database-multi-schema-orm',
-			},
-			{
-				title: 'Generate SDKs for all your APIs',
-				href: '/docs/use-cases/generate-sdks-for-all-your-apis',
-			},
-			{
-				title: 'Centralized Governance, Monitoring, Access Controls and Logging across your APIs',
-				href: '/docs/use-cases/centralized-governance-monitoring-access-controls-and-logging-across-your-apis',
-			},
-			{
-				title: 'Enabling your Organization to become API first',
-				href: '/docs/use-cases/enabling-your-organization-to-become-api-first',
 			},
 		],
 	},
@@ -957,6 +1027,10 @@ const navigation = [
 				href: '/docs/clients-reference/react-query',
 			},
 			{
+				title: 'React Relay',
+				href: '/docs/clients-reference/react-relay',
+			},
+			{
 				title: 'Solid Query',
 				href: '/docs/clients-reference/solid-query',
 			},
@@ -1059,6 +1133,10 @@ const navigation = [
 					{
 						title: 'Token-based Authentication',
 						href: '/docs/wundergraph-config-ts-reference/configure-token-based-authentication',
+					},
+					{
+						title: 'Custom Claims',
+						href: '/docs/wundergraph-config-ts-reference/configure-custom-claims',
 					},
 					{
 						title: 'Authorization / RBAC',
@@ -1362,38 +1440,6 @@ const navigation = [
 		],
 	},
 	{ type: 'divider' },
-	{
-		title: 'About WunderGraph',
-		href: '/manifesto',
-		icon: <QuestionMarkCircleIcon />,
-		links: [
-			{
-				title: 'Manifesto',
-				href: '/manifesto',
-			},
-			{
-				title: 'Usage Information',
-				href: '/usage-information',
-			},
-			{
-				title: 'Frequently Asked Questions',
-				links: [
-					{
-						title: 'How is WunderGraph faster and more secure than other GraphQL solutions?',
-						href: '/docs/frequently-asked-questions/how-is-wundergraph-faster-and-more-secure-than-other-graphql-solutions',
-					},
-					{
-						title: 'How is the developer experience different from legacy GraphQL?',
-						href: '/docs/frequently-asked-questions/how-is-the-developer-experience-different-from-legacy-graphql',
-					},
-					{
-						title: 'How is server-side only GraphQL different from client-side GraphQL?',
-						href: '/docs/frequently-asked-questions/how-is-server-side-only-graphql-different-from-client-side-graphql',
-					},
-				],
-			},
-		],
-	},
 	// {
 	// 	title: 'Troubleshooting',
 	// 	href: '/docs/troubleshooting',