+++ docs-website/src/pages/docs/getting-started/index.jsx
@@ -5,6 +5,8 @@ import { NextLogo } from '@/components/logos/next';
 import { ViteLogo } from '@/components/logos/vite';
 import { RemixLogo } from '@/components/logos/remix';
 import { SvelteLogo } from '@/components/logos/svelte';
+import { NuxtLogo } from '@/components/logos/nuxt';
+import { RelayLogo } from '@/components/logos/relay';
 
 export default function GettingStartedPage() {
 	return (
@@ -52,10 +54,23 @@ export default function GettingStartedPage() {
 						variant="plain"
 					/>
 					<QuickLink
+						title="Relay"
+						description="Learn how to use WunderGraph with Relay."
+						logo={<RelayLogo />}
+						href="/docs/getting-started/relay-quickstart"
+						variant="plain"
+					/>
+					<QuickLink
 						title="SvelteKit"
 						description="Learn how to use WunderGraph with SvelteKit."
 						logo={<SvelteLogo />}
 						href="/docs/getting-started/sveltekit-quickstart"
+					/>
+					<QuickLink
+						title="Nuxt"
+						description="Learn how to use WunderGraph with Nuxt."
+						logo={<NuxtLogo />}
+						href="/docs/getting-started/nuxt-quickstart"
 						variant="plain"
 					/>
 				</QuickLinks>