+++ packages/sdk/src/server/server.ts
@@ -61,9 +61,9 @@ if (process.env.START_HOOKS_SERVER === 'true') {
 		});
 		WG_CONFIG = JSON.parse(configContent);
 
-		if (WG_CONFIG.api && WG_CONFIG.api?.nodeOptions?.nodeUrl) {
-			const nodeUrl = resolveConfigurationVariable(WG_CONFIG.api.nodeOptions.nodeUrl);
-			clientFactory = internalClientFactory(WG_CONFIG.api.operations, nodeUrl);
+		if (WG_CONFIG.api && WG_CONFIG.api?.nodeOptions?.nodeInternalUrl) {
+			const nodeInternalURL = resolveConfigurationVariable(WG_CONFIG.api.nodeOptions.nodeInternalUrl);
+			clientFactory = internalClientFactory(WG_CONFIG.api.operations, nodeInternalURL);
 		} else {
 			throw new Error('User defined api is not set.');
 		}
@@ -119,6 +119,7 @@ const _configureWunderGraphServer = <
 		}
 
 		const tracerProvider = configureTracerProvider(WG_CONFIG.api?.nodeOptions?.openTelemetry, logger);
+		logger.info('Starting WunderGraph Server');
 
 		startServer({
 			wundergraphDir: process.env.WG_DIR_ABS!,
@@ -166,9 +167,8 @@ export const createServer = async ({
 		logger.level = resolveServerLogLevel(config.api.serverOptions.logger.level);
 	}
 
-	//const tracerProvider = configureTracerProvider(config.api?.nodeOptions?.openTelemetry, logger);
-	const nodeURL = WG_CONFIG?.api?.nodeOptions?.nodeUrl
-		? resolveConfigurationVariable(WG_CONFIG?.api?.nodeOptions?.nodeUrl)
+	const nodeInternalURL = config?.api?.nodeOptions?.nodeInternalUrl
+		? resolveConfigurationVariable(config.api.nodeOptions.nodeInternalUrl)
 		: '';
 
 	let id = 0;
@@ -316,7 +316,7 @@ export const createServer = async ({
 			wundergraphDir,
 			webhooks: config.api.webhooks,
 			internalClientFactory: clientFactory,
-			nodeURL,
+			nodeURL: nodeInternalURL,
 		});
 		fastify.log.debug('Webhooks plugin registered');
 	}
@@ -335,7 +335,7 @@ export const createServer = async ({
 			await fastify.register(FastifyFunctionsPlugin, {
 				operations: operationsConfig.typescript_operation_files,
 				internalClientFactory: clientFactory,
-				nodeURL,
+				nodeURL: nodeInternalURL,
 			});
 			fastify.log.debug('Functions plugin registered');
 		}