+++ pkg/apihandler/apihandler.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"net/http"
 	"net/http/httputil"
-	"net/url"
 	"os"
 	"strconv"
 	"strings"
@@ -77,8 +76,8 @@ type WgRequestParams struct {
 	SubsribeOnce bool
 }
 
-func NewWgRequestParams(url *url.URL) WgRequestParams {
-	q := url.Query()
+func NewWgRequestParams(r *http.Request) WgRequestParams {
+	q := r.URL.Query()
 	return WgRequestParams{
 		UseJsonPatch: q.Has(WgJsonPatchParam),
 		UseSse:       q.Has(WgSseParam),
@@ -290,7 +289,11 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 			},
 		)
 		if err != nil {
-			r.log.Error("registerS3UploadClient", zap.Error(err))
+			r.log.Info("unable to register S3 provider",
+				zap.Error(err),
+				zap.String("provider", s3Provider.Name),
+				zap.String("endpoint", loadvariable.String(s3Provider.Endpoint)),
+			)
 		} else {
 			s3Path := fmt.Sprintf("/s3/%s/upload", s3Provider.Name)
 			r.router.Handle(s3Path, http.HandlerFunc(s3.UploadFile))
@@ -749,7 +752,7 @@ type GraphQLPlaygroundHandler struct {
 }
 
 func (h *GraphQLPlaygroundHandler) ServeHTTP(w http.ResponseWriter, _ *http.Request) {
-	tpl := strings.Replace(h.html, "{{apiURL}}", h.nodeUrl, 1)
+	tpl := strings.Replace(h.html, "{{apiURL}}", h.nodeUrl, -1)
 	resp := []byte(tpl)
 
 	w.Header().Set("Content-Type", "text/html; charset=utf-8")
@@ -1451,7 +1454,7 @@ func (h *QueryHandler) handleLiveQueryEvent(ctx *resolve.Context, w http.Respons
 }
 
 func (h *QueryHandler) handleLiveQuery(r *http.Request, w http.ResponseWriter, ctx *resolve.Context, requestBuf *bytes.Buffer, flusher http.Flusher, requestLogger *zap.Logger) {
-	wgParams := NewWgRequestParams(r.URL)
+	wgParams := NewWgRequestParams(r)
 
 	done := ctx.Context.Done()
 
@@ -1726,6 +1729,14 @@ func (h *SubscriptionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) 
 			"wg.operation.type": h.operation.OperationType.String(),
 		},
 	)
+
+	// recover from panic if one occured. Set err to nil otherwise.
+	defer func() {
+		if r := recover(); r != nil {
+			requestLogger.Error("panic recovered", zap.Any("panic", r))
+			w.WriteHeader(http.StatusInternalServerError)
+		}
+	}()
 
 	if proceed := h.rbacEnforcer.Enforce(r); !proceed {
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
@@ -2242,8 +2253,6 @@ func (h *FunctionsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 	r = setOperationMetaData(r, h.operation)
 
-	isInternal := strings.HasPrefix(r.URL.Path, "/internal/")
-
 	ctx := pool.GetCtx(r, r, pool.Config{})
 	defer pool.PutCtx(ctx)
 
@@ -2267,11 +2276,7 @@ func (h *FunctionsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 			w.WriteHeader(http.StatusBadRequest)
 			return
 		}
-		if isInternal {
-			ctx.Variables, _, _, _ = jsonparser.Get(variablesBuf.Bytes(), "input")
-		} else {
-			ctx.Variables = variablesBuf.Bytes()
-		}
+		ctx.Variables = variablesBuf.Bytes()
 	}
 
 	if len(ctx.Variables) == 0 {
@@ -2391,7 +2396,7 @@ func (h *FunctionsHandler) handleRequest(ctx context.Context, w http.ResponseWri
 }
 
 func (h *FunctionsHandler) handleSubscriptionRequest(ctx context.Context, w http.ResponseWriter, r *http.Request, input []byte, requestLogger *zap.Logger) {
-	wgParams := NewWgRequestParams(r.URL)
+	wgParams := NewWgRequestParams(r)
 
 	setSubscriptionHeaders(w)
 	buf := pool.GetBytesBuffer()
@@ -2463,6 +2468,13 @@ func setOperationMetaData(r *http.Request, operation *wgpb.Operation) *http.Requ
 }
 
 func getOperationMetaData(r *http.Request) *OperationMetaData {
+	if r == nil {
+		return nil
+	}
+	ctx := r.Context()
+	if ctx == nil {
+		return nil
+	}
 	maybeMetaData := r.Context().Value("operationMetaData")
 	if maybeMetaData == nil {
 		return nil
@@ -2495,7 +2507,7 @@ func getHooksFlushWriter(ctx *resolve.Context, r *http.Request, w http.ResponseW
 }
 
 func getFlushWriter(ctx context.Context, variables []byte, r *http.Request, w http.ResponseWriter) (context.Context, *httpFlushWriter, bool) {
-	wgParams := NewWgRequestParams(r.URL)
+	wgParams := NewWgRequestParams(r)
 
 	flusher, ok := w.(http.Flusher)
 	if !ok {