+++ pkg/node/config.go
@@ -3,6 +3,7 @@ package node
 import (
 	"fmt"
 	"os"
+	"strconv"
 	"strings"
 	"time"
 
@@ -32,7 +33,7 @@ type WunderNodeConfig struct {
 	Api    *apihandler.Api
 }
 
-func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (WunderNodeConfig, error) {
+func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (*WunderNodeConfig, error) {
 	const (
 		defaultTimeout = 10 * time.Second
 	)
@@ -41,12 +42,37 @@ func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (WunderNodeConfig,
 
 	logLevel, err := logging.FindLogLevel(logLevelStr)
 	if err != nil {
-		return WunderNodeConfig{}, err
+		return nil, err
+	}
+
+	nodeHost := loadvariable.String(graphConfig.Api.NodeOptions.Listen.Host)
+
+	var nodePort uint64
+	rawNodePort := loadvariable.String(graphConfig.Api.NodeOptions.Listen.Port)
+	if rawNodePort != "" {
+		nodePort, err = strconv.ParseUint(rawNodePort, 10, 16)
+		if err != nil {
+			return nil, fmt.Errorf("can't parse node port %s: %w", rawNodePort, err)
+		}
+	}
+
+	var internalNodePort uint64
+	rawInternalNodePort := loadvariable.String(graphConfig.Api.NodeOptions.ListenInternal.Port)
+	if rawInternalNodePort != "" {
+		internalNodePort, err = strconv.ParseUint(rawInternalNodePort, 10, 16)
+		if err != nil {
+			return nil, fmt.Errorf("can't parse node internal port %s: %w", rawInternalNodePort, err)
+		}
 	}
 
 	listener := &apihandler.Listener{
-		Host: loadvariable.String(graphConfig.Api.NodeOptions.Listen.Host),
-		Port: uint16(loadvariable.Int(graphConfig.Api.NodeOptions.Listen.Port)),
+		Host: nodeHost,
+		Port: uint16(nodePort),
+	}
+
+	internalListener := &apihandler.Listener{
+		Host: nodeHost,
+		Port: uint16(internalNodePort),
 	}
 
 	telemetry := &apihandler.OpenTelemetry{
@@ -68,7 +94,7 @@ func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (WunderNodeConfig,
 		if inMemoryCacheConfig != "" {
 			cacheSize, err = units.RAMInBytes(inMemoryCacheConfig)
 			if err != nil {
-				return WunderNodeConfig{}, fmt.Errorf("can't parse %s = %q: %w", wgInMemoryCacheConfigEnvKey, inMemoryCacheConfig, err)
+				return nil, fmt.Errorf("can't parse %s = %q: %w", wgInMemoryCacheConfigEnvKey, inMemoryCacheConfig, err)
 			}
 		}
 		if cacheSize > 0 {
@@ -96,9 +122,10 @@ func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (WunderNodeConfig,
 			AuthenticationConfig:  graphConfig.Api.AuthenticationConfig,
 			Webhooks:              graphConfig.Api.Webhooks,
 			Options: &apihandler.Options{
-				ServerUrl:     strings.TrimSuffix(loadvariable.String(graphConfig.Api.ServerOptions.ServerUrl), "/"),
-				PublicNodeUrl: loadvariable.String(graphConfig.Api.NodeOptions.PublicNodeUrl),
-				Listener:      listener,
+				ServerUrl:        strings.TrimSuffix(loadvariable.String(graphConfig.Api.ServerOptions.ServerUrl), "/"),
+				PublicNodeUrl:    loadvariable.String(graphConfig.Api.NodeOptions.PublicNodeUrl),
+				Listener:         listener,
+				InternalListener: internalListener,
 				Logging: apihandler.Logging{
 					Level: logLevel,
 				},
@@ -115,5 +142,5 @@ func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (WunderNodeConfig,
 		},
 	}
 
-	return config, nil
+	return &config, nil
 }